Extracted commands into own files
