+++ lib/commands/completed.js
@@ -1,0 +1,7 @@
+exports.handler = function current(shop) {
+  var completed = shop.getData('completed')
+  if (completed)
+	completed.forEach(function (completed) {
+		console.log(shop.__('exercise.' + completed))
+	})
+}
