+++ lib/commands/current.js
@@ -1,0 +1,3 @@
+exports.handler = function current(shop) {
+  console.log(shop.__('exercise.' + shop.current))
+}
