+++ lib/commands/list.js
@@ -1,0 +1,5 @@
+exports.handler = function (shop) {
+  shop.exercises.forEach(function (name) {
+    console.log(shop.__('exercise.' + name))
+  })
+}
