+++ lib/commands/reset.js
@@ -1,0 +1,4 @@
+exports.handler = function (shop) {
+  shop.reset()
+  return console.log(shop.__('progress.reset', {title: shop.__('title')}))
+}
