+++ lib/commands/verify.js
@@ -1,0 +1,1 @@
+exports.handler = require('../runOrVerify').bind(null, 'verify')
