+++ lib/modifiers/lang.js
@@ -1,0 +1,3 @@
+exports.handler = function (shop, lang) {
+  shop.selectLanguage(lang)
+}
