+++ lib/modifiers/version.js
@@ -1,0 +1,1 @@
+exports.handler = require('../printVersion')
