+++ lib/print.js
@@ -2,6 +2,7 @@ const fs          = require('fs')
     , path        = require('path')
     , colorsTmpl  = require('colors-tmpl')
     , through     = require('through')
+    , chalk       = require('chalk')
     , msee        = require('msee')
     , mseeOptions = {
           paragraphStart: ''
@@ -101,7 +102,13 @@ function localisedFirstFileStream (appName, appDir, files, lang) {
   return file ? createFileStream(appName, appDir, file) : null
 }
 
+function error () {
+  var pr = chalk.bold.red
+  console.log(pr.apply(pr, arguments))
+  process.exit(-1)
+}
 
-module.exports.text = printText
-module.exports.localisedFileStream = localisedFileStream
-module.exports.localisedFirstFileStream = localisedFirstFileStream
+exports.error = error
+exports.text = printText
+exports.localisedFileStream = localisedFileStream
+exports.localisedFirstFileStream = localisedFirstFileStream
